+++ stable_baselines/__init__.py
@@ -2,10 +2,10 @@ from stable_baselines.a2c import A2C
 from stable_baselines.acer import ACER
 from stable_baselines.acktr import ACKTR
 from stable_baselines.ddpg import DDPG
-from stable_baselines.deepq import DeepQ
+from stable_baselines.deepq import DQN
 from stable_baselines.gail import GAIL
 from stable_baselines.ppo1 import PPO1
 from stable_baselines.ppo2 import PPO2
 from stable_baselines.trpo_mpi import TRPO
 
-__version__ = "1.0.8.rc2"
+__version__ = "2.0.0.rc0"