+++ stable_baselines/common/__init__.py
@@ -1,8 +1,9 @@
 # flake8: noqa F403
 from stable_baselines.common.console_util import fmt_row, fmt_item, colorize
 from stable_baselines.common.dataset import Dataset
-from stable_baselines.common.math_util import discount, discount_with_boundaries, explained_variance, explained_variance_2d,\
-    flatten_arrays, unflatten_vector
+from stable_baselines.common.math_util import discount, discount_with_boundaries, explained_variance, \
+    explained_variance_2d, flatten_arrays, unflatten_vector
 from stable_baselines.common.misc_util import zipsame, unpack, EzPickle, set_global_seeds, pretty_eta, RunningAvg,\
     boolean_flag, get_wrapper_by_name, relatively_safe_pickle_dump, pickle_load
-from stable_baselines.common.base_class import BaseRLModel, SetVerbosity, TensorboardWriter
+from stable_baselines.common.base_class import BaseRLModel, ActorCriticRLModel, OffPolicyRLModel, SetVerbosity, \
+    TensorboardWriter