+++ stable_baselines/ddpg/ddpg.py
@@ -11,7 +11,7 @@ import tensorflow.contrib as tc
 from mpi4py import MPI
 
 from stable_baselines import logger
-from stable_baselines.common import tf_util, BaseRLModel, SetVerbosity, TensorboardWriter
+from stable_baselines.common import tf_util, OffPolicyRLModel, SetVerbosity, TensorboardWriter
 from stable_baselines.common.vec_env import VecEnv
 from stable_baselines.common.mpi_adam import MpiAdam
 from stable_baselines.ddpg.policies import DDPGPolicy
@@ -130,7 +130,7 @@ def get_perturbed_actor_updates(actor, perturbed_actor, param_noise_stddev, verb
     return tf.group(*updates)
 
 
-class DDPG(BaseRLModel):
+class DDPG(OffPolicyRLModel):
     """
     Deep Deterministic Policy Gradient (DDPG) model
 
@@ -176,7 +176,8 @@ class DDPG(BaseRLModel):
                  render=False, render_eval=False, memory_limit=100, verbose=0, tensorboard_log=None,
                  _init_setup_model=True):
 
-        super(DDPG, self).__init__(policy=policy, env=env, verbose=verbose, policy_base=DDPGPolicy,
+        # TODO: replay_buffer refactoring
+        super(DDPG, self).__init__(policy=policy, env=env, replay_buffer=None, verbose=verbose, policy_base=DDPGPolicy,
                                    requires_vec_env=False)
 
         # Parameters.
@@ -541,10 +542,11 @@ class DDPG(BaseRLModel):
         :param compute_q: (bool) compute the critic output
         :return: ([float], float) the action and critic value
         """
-        feed_dict = {self.obs_train: [obs]}
+        obs = np.array(obs).reshape((-1,) + self.observation_space.shape)
+        feed_dict = {self.obs_train: obs}
         if self.param_noise is not None and apply_noise:
             actor_tf = self.perturbed_actor_tf
-            feed_dict[self.obs_noise] = [obs]
+            feed_dict[self.obs_noise] = obs
         else:
             actor_tf = self.actor_tf
 
@@ -930,18 +932,27 @@ class DDPG(BaseRLModel):
                                 pickle.dump(self.eval_env.get_state(), file_handler)
 
     def predict(self, observation, state=None, mask=None, deterministic=True):
-        observation = np.array(observation).reshape(self.observation_space.shape)
+        observation = np.array(observation)
+        vectorized_env = self._is_vectorized_observation(observation, self.observation_space)
 
-        action, _ = self._policy(observation, apply_noise=not deterministic, compute_q=False)
-        action = action * np.abs(self.action_space.low)  # scale the output for the prediction
-        if self._vectorize_action:
-            return np.array([action]), np.array([None])
-        else:
-            return action, None
+        observation = observation.reshape((-1,) + self.observation_space.shape)
+        actions, _, = self._policy(observation, apply_noise=not deterministic, compute_q=False)
+        actions = actions.reshape((-1,) + self.action_space.shape)  # reshape to the correct action shape
+        actions = actions * np.abs(self.action_space.low)  # scale the output for the prediction
+
+        if not vectorized_env:
+            actions = actions[0]
+
+        return actions, None
 
     def action_probability(self, observation, state=None, mask=None):
+        observation = np.array(observation)
+        vectorized_env = self._is_vectorized_observation(observation, self.observation_space)
+
+        observation = observation.reshape((-1,) + self.observation_space.shape)
+
         # here there are no action probabilities, as DDPG is continuous
-        if self._vectorize_action:
+        if vectorized_env:
             return self.sess.run(self.policy_tf.policy_proba, feed_dict={self.obs_train: observation})
         else:
             return self.sess.run(self.policy_tf.policy_proba, feed_dict={self.obs_train: observation})[0]