+++ tests/test_deterministic.py
@@ -1,6 +1,6 @@
 import pytest
 
-from stable_baselines import A2C, ACER, ACKTR, DeepQ, DDPG, PPO1, PPO2, TRPO
+from stable_baselines import A2C, ACER, ACKTR, DQN, DDPG, PPO1, PPO2, TRPO
 from stable_baselines.ddpg import AdaptiveParamNoiseSpec
 from stable_baselines.common.identity_env import IdentityEnv, IdentityEnvBox
 from stable_baselines.common.vec_env import DummyVecEnv
@@ -12,7 +12,7 @@ LEARN_FUNC_DICT = {
     'a2c': lambda e: A2C(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
     'acer': lambda e: ACER(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
     'acktr': lambda e: ACKTR(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
-    'deepq': lambda e: DeepQ(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
+    'dqn': lambda e: DQN(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
     'ddpg': lambda e: DDPG(policy="MlpPolicy", env=e, param_noise=PARAM_NOISE_DDPG).learn(total_timesteps=1000),
     'ppo1': lambda e: PPO1(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
     'ppo2': lambda e: PPO2(policy="MlpPolicy", env=e).learn(total_timesteps=1000),
@@ -21,7 +21,7 @@ LEARN_FUNC_DICT = {
 
 
 @pytest.mark.slow
-@pytest.mark.parametrize("model_name", ['a2c', 'acer', 'acktr', 'deepq', 'ppo1', 'ppo2', 'trpo'])
+@pytest.mark.parametrize("model_name", ['a2c', 'acer', 'acktr', 'dqn', 'ppo1', 'ppo2', 'trpo'])
 def test_identity(model_name):
     """
     Test if the algorithm (with a given policy)